基本完成了菜谱推荐（需后续修改）

+++ src/main/java/org/fridge/controller/MenuFavouriteController.java
@@ -30,7 +30,7 @@ public class MenuFavouriteController {
 
     @PostMapping(value = "/menu/select/favourite")
     @ResponseBody
-    public ApiResponse<List<MenuFavourite>> selectFavouriteByUserId(int id) {
+    public ApiResponse<List<MenuFavourite>> selectFavouriteByUserId(Long id) {
         return Responses.ok(menuFavouriteService.selectFavourite(id));
     }
 }
